Refine CryptoContract
